requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
bcrypt>=4.0.0
google-generativeai>=0.5.0
scipy>=1.11.0
